+++ README.md
@@ -7,7 +7,7 @@
 - Understand and process natural language queries in Korean.
 - Cache city coordinates for efficient API usage.
 - Format weather forecasts in an easily readable manner.
-
+<!-- 
 ## Building the Package
 1. Prepare for Packaging:
     - Ensure the city coordinate file is saved correctly in '~/.weather_retriever'.
@@ -22,23 +22,23 @@
     ```bash
     # pip install twine
     python -m twine upload dist/*
-    ```
+    ``` -->
 
 ## Installation
 You can install WeatherRetriever in two ways:
-1. Install via PyPI:
+<!-- 1. Install via PyPI:
     ```bash
     # https://pypi.org/project/weather_retriever/0.1.13/
     pip install weather_retriever
     ```
     - This method allows you to install the package directly from PyPI.
-2. Install from GitHub:
-    ```bash
-    git clone https://github.com/KimRass/weather_retriever
-    cd weather_retriever
-    pip install .
-    ```
-    - This method is useful if you want to access the latest code or contribute to the project.
+2. Install from GitHub: -->
+```bash
+git clone https://github.com/KimRass/weather_retriever
+cd weather_retriever
+pip install .
+```
+- This method is useful if you want to access the latest code or contribute to the project.
 
 ## Usage
 ```python
